Refactor Core. Extract DensityUnitaryMatrix, QuantumState, and SparseArray to separate files. Remove include/qx/Common.hpp. Remove src/qx/Core.cpp.
