--- include/qx/Common.hpp
@@ -1,8 +1,0 @@
-#include "qx/CompileTimeConfiguration.hpp"
-#include "qx/Utils.hpp"
-
-namespace qx {
-
-using BasisVector = utils::Bitset<config::MAX_QUBIT_NUMBER>;
-
-}
